+++ productservice/src/main/java/org/example/productservice/option/OptionJPARepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface OptionJPARepository extends JpaRepository<Option, Integer> {
+public interface OptionJPARepository extends JpaRepository<Option, Long> {
 
-    List<Option> findByProductId(int productId);
-    Optional<Option> findById(int id);
+    List<Option> findByProductId(long productId);
+    Optional<Option> findById(long id);
 }
